requirements.txt
alembic==1.14.1
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
cssmin==0.2.0
Deprecated==1.2.18
Flask==3.1.0
Flask-Assets==2.1.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsmin==3.0.1
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
ordered-set==4.1.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dotenv==1.0.1
rich==13.9.4
sh==2.2.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
utiltools==0.0.5
uWSGI==2.0.28
webassets==2.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
